requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.7.0
docopt==0.6.2
filelock==3.17.0
gunicorn==23.0.0
identify==2.6.6
idna==3.10
iniconfig==2.0.0
jsonpickle==4.0.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
pytest==8.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
virtualenv==20.29.1
yarg==0.1.10
